chore(Page) fix typos, file names, formatting
