+++ packages/react-core/src/components/Page/examples/PageHorizontalNav.tsx
@@ -2,48 +2,28 @@ import React from 'react';
 import {
   Page,
   Masthead,
-  MastheadToggle,
   MastheadMain,
   MastheadBrand,
   MastheadContent,
-  PageSidebar,
   PageSection,
   PageSectionVariants,
-  PageToggleButton,
   Toolbar,
   ToolbarContent,
   ToolbarItem
 } from '@patternfly/react-core';
-import BarsIcon from '@patternfly/react-icons/dist/esm/icons/bars-icon';
 
-export const VerticalPage: React.FunctionComponent = () => {
-  const [isNavOpen, setIsNavOpen] = React.useState(true);
-
-  const onNavToggle = () => {
-    setIsNavOpen(!isNavOpen);
-  };
-
+export const PageHorizontalNav: React.FunctionComponent = () => {
   const headerToolbar = (
-    <Toolbar id="vertical-toolbar">
+    <Toolbar id="horizontal-toolbar">
       <ToolbarContent>
+        <ToolbarItem>Navigation</ToolbarItem>
         <ToolbarItem>header-tools</ToolbarItem>
       </ToolbarContent>
     </Toolbar>
   );
 
-  const Header = (
-    <Masthead>
-      <MastheadToggle>
-        <PageToggleButton
-          variant="plain"
-          aria-label="Global navigation"
-          isNavOpen={isNavOpen}
-          onNavToggle={onNavToggle}
-          id="vertical-nav-toggle"
-        >
-          <BarsIcon />
-        </PageToggleButton>
-      </MastheadToggle>
+  const header = (
+    <Masthead inset={{ default: 'insetXs' }}>
       <MastheadMain>
         <MastheadBrand href="https://patternfly.org" target="_blank">
           Logo
@@ -52,10 +32,9 @@ export const VerticalPage: React.FunctionComponent = () => {
       <MastheadContent>{headerToolbar}</MastheadContent>
     </Masthead>
   );
-  const Sidebar = <PageSidebar nav="Navigation" isNavOpen={isNavOpen} id="vertical-sidebar" />;
 
   return (
-    <Page header={Header} sidebar={Sidebar}>
+    <Page header={header}>
       <PageSection variant={PageSectionVariants.darker}>Section with darker background</PageSection>
       <PageSection variant={PageSectionVariants.dark}>Section with dark background</PageSection>
       <PageSection variant={PageSectionVariants.light}>Section with light background</PageSection>
