+++ packages/react-core/src/components/Page/examples/PageTertiaryNav.tsx
@@ -12,7 +12,7 @@ import {
   ToolbarItem
 } from '@patternfly/react-core';
 
-export const TertiaryPage: React.FunctionComponent = () => {
+export const PageTertiaryNav: React.FunctionComponent = () => {
   const headerToolbar = (
     <Toolbar id="tertiary-toolbar">
       <ToolbarContent>
@@ -21,7 +21,7 @@ export const TertiaryPage: React.FunctionComponent = () => {
     </Toolbar>
   );
 
-  const Header = (
+  const header = (
     <Masthead display={{ default: 'stack' }} inset={{ default: 'insetXs' }}>
       <MastheadMain>
         <MastheadBrand href="https://patternfly.org" target="_blank">
@@ -33,7 +33,7 @@ export const TertiaryPage: React.FunctionComponent = () => {
   );
 
   return (
-    <Page header={Header} tertiaryNav="Navigation">
+    <Page header={header} tertiaryNav="Navigation">
       <PageSection variant={PageSectionVariants.darker}>Section with darker background</PageSection>
       <PageSection variant={PageSectionVariants.dark}>Section with dark background</PageSection>
       <PageSection variant={PageSectionVariants.light}>Section with light background</PageSection>
