+++ packages/react-core/src/components/Page/examples/PageVerticalNav.tsx
@@ -2,28 +2,48 @@ import React from 'react';
 import {
   Page,
   Masthead,
+  MastheadToggle,
   MastheadMain,
   MastheadBrand,
   MastheadContent,
+  PageSidebar,
   PageSection,
   PageSectionVariants,
+  PageToggleButton,
   Toolbar,
   ToolbarContent,
   ToolbarItem
 } from '@patternfly/react-core';
+import BarsIcon from '@patternfly/react-icons/dist/esm/icons/bars-icon';
 
-export const HorizontalPage: React.FunctionComponent = () => {
+export const PageVerticalNav: React.FunctionComponent = () => {
+  const [isNavOpen, setIsNavOpen] = React.useState(true);
+
+  const onNavToggle = () => {
+    setIsNavOpen(!isNavOpen);
+  };
+
   const headerToolbar = (
-    <Toolbar id="horizontal-toolbar">
+    <Toolbar id="vertical-toolbar">
       <ToolbarContent>
-        <ToolbarItem>Navigation</ToolbarItem>
         <ToolbarItem>header-tools</ToolbarItem>
       </ToolbarContent>
     </Toolbar>
   );
 
-  const Header = (
-    <Masthead inset={{ default: 'insetXs' }}>
+  const header = (
+    <Masthead>
+      <MastheadToggle>
+        <PageToggleButton
+          variant="plain"
+          aria-label="Global navigation"
+          isNavOpen={isNavOpen}
+          onNavToggle={onNavToggle}
+          id="vertical-nav-toggle"
+        >
+          <BarsIcon />
+        </PageToggleButton>
+      </MastheadToggle>
       <MastheadMain>
         <MastheadBrand href="https://patternfly.org" target="_blank">
           Logo
@@ -32,8 +52,11 @@ export const HorizontalPage: React.FunctionComponent = () => {
       <MastheadContent>{headerToolbar}</MastheadContent>
     </Masthead>
   );
+
+  const sidebar = <PageSidebar nav="Navigation" isNavOpen={isNavOpen} id="vertical-sidebar" />;
+
   return (
-    <Page header={Header}>
+    <Page header={header} sidebar={sidebar}>
       <PageSection variant={PageSectionVariants.darker}>Section with darker background</PageSection>
       <PageSection variant={PageSectionVariants.dark}>Section with dark background</PageSection>
       <PageSection variant={PageSectionVariants.light}>Section with light background</PageSection>
